// apps/api/prisma/seed/generated/graphql.ts
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: any
}
// ...
export enum BillboardStatusType {
  Approved = 'APPROVED',
  InstallationInprogress = 'INSTALLATION_INPROGRESS',
  Live = 'LIVE',
  New = 'NEW',
  OnHold = 'ON_HOLD',
  Rejected = 'REJECTED',
  Verified = 'VERIFIED',
}
// ...
export enum BillboardType {
  Classic = 'CLASSIC',
  Led = 'LED',
  Neon = 'NEON',
  ThreeDimensional = 'THREE_DIMENSIONAL',
  Vinyl = 'VINYL',
}
// ...
export type CreateBillboardInput = {
  address: Scalars['String']
  angle: Scalars['Int']
  elevation: Scalars['Int']
  height: Scalars['Int']
  images: Array<Scalars['String']>
  impressionsPerDay: Scalars['Int']
  lat: Scalars['Float']
  lng: Scalars['Float']
  minBookingDays: Scalars['Int']
  ownerId: Scalars['String']
  type: BillboardType
  width: Scalars['Int']
}

export type CreateBillboardStatusInput = {
  agentId: Scalars['String']
  billboardId: Scalars['Int']
  notes: Scalars['String']
  status: BillboardStatusType
}

export type CreateBookingInput = {
  billboardId: Scalars['Int']
  campaignId: Scalars['Int']
  pricePerDay: Scalars['Int']
}

export type CreateBookingWithoutCampaignIdInput = {
  billboardId: Scalars['Int']
  pricePerDay: Scalars['Int']
}

export type CreateCampaignInput = {
  advertiserId: Scalars['String']
  bookings: Array<CreateBookingWithoutCampaignIdInput>
  endDate: Scalars['DateTime']
  name: Scalars['String']
  startDate: Scalars['DateTime']
}

export type CreateCampaignStatusInput = {
  agentId: Scalars['String']
  campaignId: Scalars['Int']
  status: CampaignStatusType
}

export type CreateFavoriteInput = {
  advertiserId: Scalars['String']
  billboardId: Scalars['Int']
}